Move dashboard components into their own folder
