+++ src/components/Dashboard/Mod.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Draggable } from "react-beautiful-dnd";
-import "../css/dashboard/Mod.css";
+import "../../css/dashboard/Mod.css";
 
 function Mod({index, modData, handleModRightClick, modColor}) {
     const handleRightClick = (event) => {
